+++ app/src/main/java/ids/univpm/breakout/model/Mappa.java
@@ -1,8 +1,8 @@
 package ids.univpm.breakout.model;
 
 public class Mappa {
-    private long ID_mappa;
-    private long ID_piano;
+    private Integer ID_mappa;
+    private Integer ID_piano;
 //    private Grafo grafo;
     private String nome;
     private String urlImmagine;
@@ -26,19 +26,19 @@ public class Mappa {
         this.urlImmagine = urlImmagine;
     }
 
-    public long getID_mappa() {
+    public Integer getID_mappa() {
         return ID_mappa;
     }
 
-    public void setID_mappa(long ID_mappa) {
+    public void setID_mappa(Integer ID_mappa) {
         this.ID_mappa = ID_mappa;
     }
 
-    public long getID_piano() {
+    public Integer getID_piano() {
         return ID_piano;
     }
 
-    public void setID_piano(long ID_piano) {
+    public void setID_piano(Integer ID_piano) {
         this.ID_piano = ID_piano;
     }
