+++ app/src/main/java/ids/univpm/breakout/model/Tronco.java
@@ -2,17 +2,17 @@ package ids.univpm.breakout.model;
 
 
 public class Tronco extends Collegamento{
-    private long ID_mappa;
+    private Integer ID_mappa;
 
 
     public Tronco(){
     };
 
-    public long getID_mappa() {
+    public Integer getID_mappa() {
         return ID_mappa;
     }
 
-    public void setID_mappa(long ID_mappa) {
+    public void setID_mappa(Integer ID_mappa) {
         this.ID_mappa = ID_mappa;
     }
 }
